Fix AudioBitmapAnimation hang at end of playback
Add flag AudioPlayer_CurrentTime_Unavailable and set it to true when
MPlayer returns an error for the current time.  Check if this flag is
true when determining if the animation has finished.  This fixes the
animation hanging at the end of playback when track position does not
reach the end of available frames.

Add flag AudioPlayer_CurrentTime_Pending to keep track of whether or
not a request for playback time has been sent to MPlayer.  This fixes
queuing up tens of requests when MPlayer is loading.

In the future, this entire class should be rewritten to integrate
with a cross-platform audio library.  But for now, classic MPlayer
lives another day.