Layer: Add GetEnumerator
Add GetEnumerator to Layer that returns an IEnumerable of Color
representing all the pixels in the given layer.

This makes "foreach" work on pixel values, albeit indirectly, e.g.
> foreach (var item in sampleLayer.GetEnumerator ()) { ... }

Add more unit tests!

NOTE:  NUnit favors enumerator-based comparison, so defining the
collection as IEnumerable requires overriding GetHashCode() and a
bunch of other steps that don't really make sense for this...

See https://stackoverflow.com/questions/10737209/assert-areequal-does-not-use-my-equals-overrides-on-an-ienumerable-implementati
And https://stackoverflow.com/questions/2860484/c-ienumerable-getenumerator-a-simple-simple-example-please