output: Switch to DeviceConfiguration, drop sleep
Switch to DeviceConfiguration and ReadOnlyDeviceConfiguration for
output devices.  This offers better organization and paves the way
for animations that scale to any update rate and LED count.

Remove SleepForAnimation and related code.  Animations will now run
as fast as possible in preparation for timing-independent animations.
In the meantime, this doesn't break anything for the G35 lights,
either - most sleeps were 0 to 1 ms long.