utility: Add IntFraction, tracks between ints
Add IntFraction class, providing a way to track fractional progress
towards the next integer value.  This acts as a convenience wrapper
to storing a double and manually truncating and subtracting with each
need of the integer value.

For example...

> var shiftAmount = new IntFraction(0);
> shiftAmount += 1.5;
> movePixels(shiftAmount.TakeInt());
[shifts by 1 pixel]
> shiftAmount += 1.5;
> movePixels(shiftAmount.TakeInt());
[shifts by 2 pixels]
> movePixels(shiftAmount.TakeInt());
[shifts by 0 pixels]

...would move pixels in integer steps, first 1 (leaving 0.5 inside),
then 2 (leaving 0 inside), then 0 (as there's nothing inside).

Add unit tests.