main: Scale transition to FPS, fix conflicts
Adjust the rate of transitions between colors according to the device
configuration, adding more intermediate frames when the output device
updates faster.

Add LayerTransformer to AddToAnimQueue, which allows reordering
transformations so that interrupting the queued animation no longer
results in broken transitions.  No matter when the commands are sent,
colors and brightness values will all correctly sum.

This removes the need for "!" to override the warning.

Clean up FillLights_* functions as they're no longer needed and don't
integrate well with the delegate-based AddToAnimQueue.