queue: Fix LED_Queue locking on PushToQueue(bool)
Fix LED_Queue locking when pushing to the queue without specifying a
frame.  Between checking if the queue is empty and trying to add to
the queue, the queue could be emptied.

This wasn't revealed until the combination of the new queue-stealing
AddToAnimQueue method that correctly interrupts fades, and the faster
LED set that results in pushing to queue more often.