output: Fix Arduino null reference racing
Fix racing between system initialization, shutdown, and sending data
to the Arduino output by locking on any access to the serial device.
This fixes issues with reconnecting to a disconnected device while
the main queue is running.

Add an UnlockedShutdownSystem() for use within internal functions
that properly lock the serial device.

Also fix two invalid uses of deviceConfig when it was null.