main: Cleanup resetting of output, fix crashes
Cleanup resetting of the output system, introducing a
System.Threading.ManualResetEvent to track when the output system is
ready.

Bail out from sending to a non-ready output system in a few more
places.

And finally, drop the keyboard input discarding loop.  That's not
needed, and can be frustrating to deal with.