anim: Fix BeatPulse, RaveMood to use DeviceConfig
Port BeatPulse and RaveMood reactive animations to use the device
configuration for scaling according to light count, strand length,
and update delay (refresh rate).

Add more fading to strobe effects in RaveMood, tempering the
harshness of the strobing.  Sorta.

This isn't an exact port; the look-'n-feel of the animations have
changed - hopefully for the better!  Still tweaking.

(And it only took me over 4 years to get to this point!)