utility: Add ScaledAverage, EMA scaled to FPS
Add ScaledAverage class, providing an Exponential Moving Average that
adjusts the percentage according to the output device update rate.
This makes it easier to apply a consistent level of smoothing in
situations where one cannot interpolate to a well-defined ending
point.

For example...

> var sampleDevice = new DeviceConfiguration (1, 1);
> var sampleDeviceRO =
>     new ReadOnlyDeviceConfiguration (sampleDevice);
> var sampleFilter = new ScaledAverage (sampleDeviceRO);
>
> sampleFilter.TimeConstant = 300; // ms
> sampleDevice.SetUpdateRate (50); // ms
> Console.WriteLine (sampleFilter.WeightMultiplier);
[prints "0.285714286"]
> sampleDevice.SetUpdateRate (4);  // ms
> Console.WriteLine (sampleFilter.WeightMultiplier);
[prints "0.026315789"]

...the much lower percentage for the average in the second case would
result in a longer time for the filter to reach the new value,
accommodating the 12.5x faster update rate.

Add unit tests.  So many tests.