color: Tweak ambient color to feel more "natural"
Tweak the ambient color based on observation with newer LED sets,
e.g. the Adafruit DotStar series.  This matches the 5 pm daytime
color in IntervalAnimation, providing a slightly more natural, warm
color, deemphasizing green and adding blue.

As expected, this is a subjective change.  Feel free to raise
complaints in the form of issues.